+++ tests/lightning/dataset_ng/test_data_module_zarr.py
@@ -1,42 +1,8 @@
-import numpy as np
-import pytest
 import zarr
 from numpy import array_equal
 
 from careamics.config.data import NGDataConfig
 from careamics.lightning.dataset_ng.data_module import CareamicsDataModule
-
-
-@pytest.fixture
-def zarr_with_target_and_mask(tmp_path) -> str:
-    """Three arrays in a single zarr group at the root."""
-    path = tmp_path / "zarr_with_target_and_mask.zarr"
-    zarr_file = zarr.create_group(path)
-    assert path.exists()
-
-    # arrays
-    arrays = np.arange(3 * 16 * 16).reshape((3, 16, 16))
-    targets = np.arange(3 * 16 * 16).reshape((3, 16, 16))
-    masks = np.ones((3, 16, 16), dtype=bool)
-    val = np.ones((16, 16))
-
-    # exclude central frame
-    masks[1] = np.zeros((16, 16), dtype=bool)
-
-    # create groups
-    input_group = zarr_file.create_group("input")
-    target_group = zarr_file.create_group("target")
-    mask_group = zarr_file.create_group("mask")
-    val_group = zarr_file.create_group("val")
-
-    # write arrays to zarr
-    for i in range(arrays.shape[0]):
-        input_group.create_array(f"array{i}", data=arrays[i], chunks=(8, 8))
-        target_group.create_array(f"array{i}", data=targets[i], chunks=(8, 8))
-        mask_group.create_array(f"array{i}", data=masks[i], chunks=(8, 8))
-    val_group.create_array("val_array", data=val, chunks=(8, 8))
-
-    return path
 
 
 def test_zarr_data_module(zarr_with_target_and_mask):
